Save login time and trigger postLogin hook in success handler
